refactor: Don't expose ops structs to users directly
