+++ src/layers/complete_reader.rs
@@ -21,6 +21,7 @@ use std::task::Poll;
 
 use async_trait::async_trait;
 
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
