+++ src/layers/immutable_index.rs
@@ -20,6 +20,7 @@ use std::mem;
 use async_trait::async_trait;
 
 use crate::object::*;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
