+++ src/layers/tracing.rs
@@ -25,6 +25,7 @@ use futures::AsyncRead;
 use futures::FutureExt;
 use tracing::Span;
 
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
