+++ src/layers/type_eraser.rs
@@ -17,6 +17,7 @@ use std::fmt::Formatter;
 
 use async_trait::async_trait;
 
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
