+++ src/raw/accessor.rs
@@ -19,6 +19,7 @@ use async_trait::async_trait;
 use flagset::flags;
 use flagset::FlagSet;
 
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
