+++ src/raw/adapters/kv/backend.rs
@@ -16,6 +16,7 @@ use async_trait::async_trait;
 use futures::AsyncReadExt;
 
 use super::Adapter;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
