+++ src/raw/io/output/into_reader/by_range.rs
@@ -27,6 +27,7 @@ use futures::future::BoxFuture;
 use futures::ready;
 use tokio::io::ReadBuf;
 
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
