+++ src/raw/io/walk.rs
@@ -18,6 +18,7 @@ use std::mem;
 use async_trait::async_trait;
 
 use crate::object::*;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
