+++ src/raw/layer.rs
@@ -16,6 +16,7 @@ use std::fmt::Debug;
 
 use async_trait::async_trait;
 
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
 
@@ -42,6 +43,7 @@ use crate::*;
 /// use std::sync::Arc;
 ///
 /// use async_trait::async_trait;
+/// use opendal::ops::*;
 /// use opendal::raw::*;
 /// use opendal::*;
 ///
