+++ src/services/azdfs/backend.rs
@@ -31,6 +31,7 @@ use reqsign::AzureStorageSigner;
 use super::dir_stream::DirStream;
 use super::error::parse_error;
 use crate::object::ObjectMetadata;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
