+++ src/services/ghac/backend.rs
@@ -31,6 +31,7 @@ use serde::Deserialize;
 use serde::Serialize;
 
 use super::error::parse_error;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
