+++ src/services/http/backend.rs
@@ -23,6 +23,7 @@ use http::StatusCode;
 use log::debug;
 
 use super::error::parse_error;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
