+++ src/services/ipfs/backend.rs
@@ -26,6 +26,7 @@ use prost::Message;
 
 use super::error::parse_error;
 use super::ipld::PBNode;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
