+++ src/services/obs/backend.rs
@@ -29,6 +29,7 @@ use reqsign::HuaweicloudObsSigner;
 
 use super::dir_stream::DirStream;
 use super::error::parse_error;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
