+++ src/services/oss/backend.rs
@@ -31,6 +31,7 @@ use reqsign::AliyunOssSigner;
 
 use super::dir_stream::DirStream;
 use super::error::parse_error;
+use crate::ops::*;
 use crate::raw::*;
 use crate::*;
